Upgrade to Porter secret protocol v2
This updates the plugin to work with v2 of the secrets protocol, which
adds support for context.Context and persisting secrets.

While I was making these changes, I was having a hard time getting the
paths to work (with respect to the data key). I have created a new
struct Secret that manages where data should be read from and written
to.

* If the secret value itself uses subpaths, the subpath is used to resolve the secret location. The final path segment is the secret key.

  For example, if the path_prefix is myorg/myapp, and the parameter set's secret value is set to v1/mysql/connstr, the secret path is myorg/data/myapp/v1/mysql and the plugin returns the secret's connstr.

* When writing secrets, we apply the PorterSecret configuration value as a suffix to the secret path.

  For example, if the path prefix is myorg/myapp, and we are saving a key named v1/mysql/connstr, the secret path is myorg/data/myapp/v1/mysql/PORTER_SECRET. If PorterSecret is not configured, we default it to "porter".

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>